pyproject.toml
[project]
name = "everything-about-agent"
version = "3.0.0"
description = "Agent Learning System with Multi-Backend Architecture"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-openai>=0.2.0",
    "ollama>=0.4.0",
    "requests>=2.32.0",
    "aiohttp>=3.10.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0.0",
    "loguru>=0.7.0",
    "gradio>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
